+++ tests/versions/test_v1_6_2.py
@@ -1,0 +1,17 @@
+from typing import Union
+
+from ididi import Graph, Ignore
+
+
+class MyConfig: ...
+
+def myfact(a: Ignore[Union[int, str, None]] = None) -> MyConfig:
+    return MyConfig()
+
+
+def test_resolve():
+    graph = Graph()
+    graph.node(myfact)
+    node = graph.analyze(myfact)
+    res = graph.resolve(MyConfig)
+    assert isinstance(res, MyConfig)